requirement.txt
Django==1.10.5
djangorestframework==3.5.3
django-cors-headers==2.0.0
django-extensions==1.7.5
psycopg2==2.6.2
requests==2.12.4
